requirements.txt
# Core dependencies
beautifulsoup4>=4.12.0
requests>=2.31.0
schedule>=1.2.0
tqdm>=4.66.0

# YouTube API
google-api-python-client>=2.160.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Web Scraping
selenium>=4.17.0
playwright>=1.41.0
webdriver_manager>=4.0.0

# Media Processing
ffmpeg-python>=0.2.0
pillow>=10.2.0

# Environment and Utils
python-dotenv>=1.0.0
pytz>=2024.1
tenacity>=8.2.0  # For retries
psutil>=5.9.0    # For system monitoring

# Logging and Monitoring
structlog>=24.1.0
python-json-logger>=2.0.7